+++ packages/plugin-package-build/src/index.ts
@@ -1,2 +1,4 @@
 export {packageBuild} from './plugin-package-build';
 export type {PackageBuildOptions} from './plugin-package-build';
+
+export {rollupPlugins} from '@sewing-kit/plugin-rollup';