+++ packages/plugin-package-build/CHANGELOG.md
@@ -10,6 +10,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Fixed bin file generation to ensure it always points at the `build/cjs` folder [[#178](https://github.com/Shopify/sewing-kit-next/pull/178)]
+- Reexport `rollupPlugins` from `@sewing-kit/plugin-rollup` to ease build customisation [[#179](https://github.com/Shopify/sewing-kit-next/pull/179)]
 
 ## [0.2.2] - 2021-05-28